refactor(code-builder): move collision check function to utils
